% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felm.R
\name{check_redundant_fe}
\alias{check_redundant_fe}
\title{Check if formula contains redundant FEs}
\usage{
check_redundant_fe(formula)
}
\arguments{
\item{formula}{a formula}
}
\description{
Print warning when there is something like fe1+fe1:fe2
}
\examples{
check_redundant_fe(y~ x)
check_redundant_fe(y~ x |fe1 +fe2)
check_redundant_fe(y~ x |fe1 +fe2:fe1)
check_redundant_fe(y~ x |fe2*fe1)
}
\author{
Grant McDermott, small adaptation by Matthieu Stigler
}
