% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feglm.R
\name{fepois}
\alias{fepois}
\title{Fixed-Effects Poisson Pseudo Maximum Likelihood (PPML)}
\usage{
fepois(
  formula,
  data,
  offset = NULL,
  subset = NULL,
  robust = TRUE,
  cluster = NULL,
  pseudo_r2 = TRUE,
  tol = 1e-10
)
}
\arguments{
\item{formula}{an object of class '"formula"' (or one that can be coerced to
that class): a symbolic description of the model to be fitted. Similarly to
'lm'.  See Details.}

\item{data}{a data frame containing the variables of the model.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{robust}{logical value to return a robust standard error computation.}

\item{cluster}{optional variable to group by and compute sandwich-type
robust standard errors. Should be a formula of the form \code{~ x_j} or
an object that be coerced to a formula.}

\item{tol}{tolerance value for GLM convergence criteria.}

\item{pseudo_rq}{logical value to return a correlation based metric
equivalent to a pseudo-R2.}
}
\description{
Fixed-Effects Poisson Pseudo Maximum Likelihood (PPML)
}
\seealso{
felm
}
