% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDmatrix}
\alias{makeDmatrix}
\title{Make sparse matrix of dummies from factor list}
\usage{
makeDmatrix(fl, weights = NULL)
}
\arguments{
\item{fl}{list of factors.}

\item{weights}{numeric vector. Multiplied into the rows.}
}
\value{
Returns a sparse matrix.
}
\description{
Given a list of factors, return the matrix of dummies as a sparse matrix.
}
\details{
The function returns the model matrix for a list of factors. This matrix is
not used internally by the package, but it's used in some of the
documentation for illustrative purposes.
}
\examples{

  fl <- lapply(1:3, function(i) factor(sample(3,10,replace=TRUE)))
  fl
  makeDmatrix(fl, weights=seq(0.1,1,0.1))

}

