% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sargan}
\alias{sargan}
\title{Compute Sargan's S}
\usage{
sargan(object, ..., lhs = object$lhs[1])
}
\arguments{
\item{object}{and object type '"felm"', the return value from \code{\link{felm}}.}

\item{...}{Not used at the moment.}

\item{lhs}{in case of multiple left hand sides, specify the name of the left 
hand side for which you want to compute Sargan's S.}
}
\value{
\code{sargan} returns a numeric, the Sargan's S. The Basmann statistic is
returned in the '"basmann"' attribute.
}
\description{
Compute Sargan's S
}
